Fix all linter issues

+++ lib/numbers/widgets/numbers.dart
@@ -1,18 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-
-import '../bloc/numbers_bloc.dart';
-import 'widgets.dart';
+import 'package:major_system/numbers/bloc/numbers_bloc.dart';
+import 'package:major_system/numbers/widgets/widgets.dart';
 
 class Numbers extends StatelessWidget {
-  Numbers({Key? key}) : super(key: key);
+  const Numbers({super.key});
 
   @override
   Widget build(BuildContext context) {
     return BlocBuilder<NumbersBloc, NumbersState>(
       builder: (context, state) {
         if (state is NumbersLoading) {
-          return LoadingIndicator();
+          return const LoadingIndicator();
         } else if (state is NumbersLoaded) {
           final numbers = state.numbers;
           return ListView.builder(
@@ -20,10 +19,29 @@ class Numbers extends StatelessWidget {
             itemBuilder: (context, index) {
               final number = numbers[index];
               return NumberItem(
-                  number: number,
-                  onDismissed: (direction) {
-                    BlocProvider.of<NumbersBloc>(context)
-                        .add(DeleteNumber(number));
+                number: number,
+                onDismissed: (direction) {
+                  BlocProvider.of<NumbersBloc>(context)
+                      .add(DeleteNumber(number));
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    DeleteNumberSnackBar(
+                      number: number,
+                      onUndo: () => BlocProvider.of<NumbersBloc>(context)
+                          .add(AddNumber(number)),
+                    ),
+                  );
+                },
+                onTap: () async {
+                  final removedNumber = await Navigator.of(context).push(
+                    MaterialPageRoute<dynamic>(
+                      builder: (_) {
+                        // TODO(jjochen): Implement DetailsScreen.
+                        return Container();
+                        //return DetailsScreen(id: number.id);
+                      },
+                    ),
+                  );
+                  if (removedNumber != null && context.mounted) {
                     ScaffoldMessenger.of(context).showSnackBar(
                       DeleteNumberSnackBar(
                         number: number,
@@ -31,25 +49,9 @@ class Numbers extends StatelessWidget {
                             .add(AddNumber(number)),
                       ),
                     );
-                  },
-                  onTap: () async {
-                    final removedNumber = await Navigator.of(context).push(
-                      MaterialPageRoute(builder: (_) {
-                        // TODO
-                        return Container();
-                        //return DetailsScreen(id: number.id);
-                      }),
-                    );
-                    if (removedNumber != null) {
-                      ScaffoldMessenger.of(context).showSnackBar(
-                        DeleteNumberSnackBar(
-                          number: number,
-                          onUndo: () => BlocProvider.of<NumbersBloc>(context)
-                              .add(AddNumber(number)),
-                        ),
-                      );
-                    }
-                  });
+                  }
+                },
+              );
             },
           );
         } else {
